fastapi_service/requirements.txt
aiosqlite==0.17.0
alembic==1.13.0
annotated-types==0.6.0
argon2==0.1.10
argon2-cffi==16.0.0
bcrypt==4.1.2
certifi==2023.11.17
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
dnspython==2.4.2
email-validator==2.0.0.post2
fastapi-users==12.1.2
greenlet==3.0.2
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
makefun==1.15.2
Mako==1.3.0
orjson==3.9.10
passlib==1.7.4
psycopg==3.1.16
psycopg-binary==3.1.16
psycopg-pool==3.2.0
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.5.2
pydantic-extra-types==2.2.0
pydantic-settings==2.1.0
pydantic_core==2.14.5
PyJWT==2.8.0
pypika-tortoise==0.1.6
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
setuptools==68.2.2
sniffio 
SQLAlchemy 
starlette
tortoise-orm==0.20.0
typing==3.7.4.3
typing_extensions 
ujson==5.9.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
wheel==0.41.2
zipp
pymongo
motor